chore: remove unwrap&split literal
