+++ src/flow/Cargo.toml
@@ -40,5 +40,4 @@ prost.workspace = true
 query.workspace = true
 serde_json = "1.0"
 session.workspace = true
-substrait.workspace = true
 table.workspace = true
