+++ src/flow/src/transform.rs
@@ -36,9 +36,10 @@ use substrait::substrait_proto::proto::expression::field_reference::ReferenceTyp
 use substrait::substrait_proto::proto::expression::literal::LiteralType;
 use substrait::substrait_proto::proto::expression::reference_segment::ReferenceType::StructField;
 use substrait::substrait_proto::proto::expression::{
-    Literal, MaskExpression, RexType, ScalarFunction,
+    IfThen, Literal, MaskExpression, RexType, ScalarFunction,
 };
 use substrait::substrait_proto::proto::extensions::simple_extension_declaration::MappingType;
+use substrait::substrait_proto::proto::extensions::SimpleExtensionDeclaration;
 use substrait::substrait_proto::proto::function_argument::ArgType;
 use substrait::substrait_proto::proto::r#type::Kind;
 use substrait::substrait_proto::proto::read_rel::ReadType;
@@ -74,9 +75,48 @@ macro_rules! plan_err {
     };
 }
 
+mod literal;
+use literal::{from_substrait_literal, from_substrait_type};
+
+/// In Substrait, a function can be define by an u32 anchor, and the anchor can be mapped to a name
+///
+/// So in substrait plan, a ref to a function can be a single u32 anchor instead of a full name in string
+pub struct FunctionExtensions {
+    anchor_to_name: HashMap<u32, String>,
+}
+
+impl FunctionExtensions {
+    /// Create a new FunctionExtensions from a list of SimpleExtensionDeclaration
+    pub fn try_from_proto(extensions: &[SimpleExtensionDeclaration]) -> Result<Self, Error> {
+        let mut anchor_to_name = HashMap::new();
+        for e in extensions {
+            match &e.mapping_type {
+                Some(ext) => match ext {
+                    MappingType::ExtensionFunction(ext_f) => {
+                        anchor_to_name.insert(ext_f.function_anchor, ext_f.name.clone());
+                    }
+                    _ => not_impl_err!("Extension type not supported: {ext:?}")?,
+                },
+                None => not_impl_err!("Cannot parse empty extension")?,
+            }
+        }
+        Ok(Self { anchor_to_name })
+    }
+
+    /// Get the name of a function by it's anchor
+    pub fn get(&self, anchor: &u32) -> Option<&String> {
+        self.anchor_to_name.get(anchor)
+    }
+}
+
+/// A context that holds the information of the dataflow
 pub struct DataflowContext {
+    /// `id` refer to any source table in the dataflow, and `name` is the name of the table
+    /// which is a `Vec<String>` in substrait
     id_to_name: HashMap<GlobalId, Vec<String>>,
+    /// see `id_to_name`
     name_to_id: HashMap<Vec<String>, GlobalId>,
+    /// the schema of the table
     schema: HashMap<GlobalId, RelationType>,
 }
 
@@ -116,19 +156,10 @@ impl DataflowContext {
     }
 }
 
+/// Convert Substrait Plan into Flow's TypedPlan
 pub fn from_substrait_plan(ctx: &mut DataflowContext, plan: &SubPlan) -> Result<TypedPlan, Error> {
     // Register function extension
-    let function_extension = plan
-        .extensions
-        .iter()
-        .map(|e| match &e.mapping_type {
-            Some(ext) => match ext {
-                MappingType::ExtensionFunction(ext_f) => Ok((ext_f.function_anchor, &ext_f.name)),
-                _ => not_impl_err!("Extension type not supported: {ext:?}"),
-            },
-            None => not_impl_err!("Cannot parse empty extension"),
-        })
-        .collect::<Result<HashMap<_, _>, Error>>()?;
+    let function_extension = FunctionExtensions::try_from_proto(&plan.extensions)?;
 
     // Parse relations
     match plan.relations.len() {
@@ -139,7 +170,10 @@ pub fn from_substrait_plan(ctx: &mut DataflowContext, plan: &SubPlan) -> Result<
                         Ok(from_substrait_rel(ctx, rel, &function_extension)?)
                     },
                     plan_rel::RelType::Root(root) => {
-                        Ok(from_substrait_rel(ctx, root.input.as_ref().unwrap(), &function_extension)?)
+                        let input = root.input.as_ref().with_context(|| InvalidQuerySnafu {
+                            reason: "Root relation without input",
+                        })?;
+                        Ok(from_substrait_rel(ctx, input, &function_extension)?)
                     }
                 },
                 None => plan_err!("Cannot parse plan relation: None")
@@ -152,11 +186,12 @@ pub fn from_substrait_plan(ctx: &mut DataflowContext, plan: &SubPlan) -> Result<
     }
 }
 
+/// Convert Substrait Rel into Flow's TypedPlan
 /// TODO: SELECT DISTINCT(does it get compile with something else?)
 pub fn from_substrait_rel(
     ctx: &mut DataflowContext,
     rel: &Rel,
-    extensions: &HashMap<u32, &String>,
+    extensions: &FunctionExtensions,
 ) -> Result<TypedPlan, Error> {
     match &rel.rel_type {
         Some(RelType::Project(p)) => {
@@ -242,7 +277,7 @@ fn from_substrait_agg_grouping(
     ctx: &mut DataflowContext,
     groupings: &[Grouping],
     typ: &RelationType,
-    extensions: &HashMap<u32, &String>,
+    extensions: &FunctionExtensions,
 ) -> Result<Vec<(ScalarExpr, ColumnType)>, Error> {
     let _ = ctx;
     let mut group_expr = vec![];
@@ -266,7 +301,7 @@ fn from_substrait_agg_measures(
     ctx: &mut DataflowContext,
     measures: &[Measure],
     typ: &RelationType,
-    extensions: &HashMap<u32, &String>,
+    extensions: &FunctionExtensions,
 ) -> Result<Vec<AggregateExpr>, Error> {
     let _ = ctx;
     let mut aggr_exprs = vec![];
@@ -344,7 +379,7 @@ fn from_substrait_gen_key_val_plan(
 fn from_substrait_agg_rel(
     ctx: &mut DataflowContext,
     agg: &proto::AggregateRel,
-    extensions: &HashMap<u32, &String>,
+    extensions: &FunctionExtensions,
 ) -> Result<TypedPlan, Error> {
     let input = if let Some(input) = agg.input.as_ref() {
         from_substrait_rel(ctx, input, extensions)?
@@ -408,10 +443,11 @@ fn from_substrait_agg_rel(
     })
 }
 
+/// Convert AggregateFunction into Flow's AggregateExpr
 pub fn from_substrait_agg_func(
     f: &proto::AggregateFunction,
     input_schema: &RelationType,
-    extensions: &HashMap<u32, &String>,
+    extensions: &FunctionExtensions,
     filter: Option<(ScalarExpr, ColumnType)>,
     order_by: Option<Vec<(ScalarExpr, ColumnType)>>,
     distinct: bool,
@@ -454,7 +490,7 @@ pub fn from_substrait_agg_func(
 pub fn from_substrait_scalar_func(
     f: &ScalarFunction,
     input_schema: &RelationType,
-    extensions: &HashMap<u32, &String>,
+    extensions: &FunctionExtensions,
 ) -> Result<(ScalarExpr, ColumnType), Error> {
     let fn_name = extensions.get(&f.function_reference).ok_or_else(|| {
         NotImplementedSnafu {
@@ -558,11 +594,65 @@ pub fn from_substrait_scalar_func(
     }
 }
 
+/// Convert IfThen into Flow's ScalarExpr
+pub fn from_substrait_ifthen_rex(
+    if_then: &IfThen,
+    input_schema: &RelationType,
+    extensions: &FunctionExtensions,
+) -> Result<(ScalarExpr, ColumnType), Error> {
+    let ifs: Vec<_> = if_then
+        .ifs
+        .iter()
+        .map(|if_clause| {
+            let proto_if = if_clause.r#if.as_ref().with_context(|| InvalidQuerySnafu {
+                reason: "IfThen clause without if",
+            })?;
+            let proto_then = if_clause.then.as_ref().with_context(|| InvalidQuerySnafu {
+                reason: "IfThen clause without then",
+            })?;
+            let cond = from_substrait_rex(proto_if, input_schema, extensions)?;
+            let then = from_substrait_rex(proto_then, input_schema, extensions)?;
+            Ok((cond, then))
+        })
+        .try_collect()?;
+    // if no else is presented
+    let els = if_then
+        .r#else
+        .as_ref()
+        .map(|e| from_substrait_rex(e, input_schema, extensions))
+        .transpose()?
+        .unwrap_or_else(|| {
+            (
+                ScalarExpr::literal_null(),
+                ColumnType::new_nullable(CDT::null_datatype()),
+            )
+        });
+    fn build_if_then(
+        mut next_if_then: impl Iterator<Item = ((ScalarExpr, ColumnType), (ScalarExpr, ColumnType))>,
+        els: (ScalarExpr, ColumnType),
+    ) -> (ScalarExpr, ColumnType) {
+        if let Some((cond, then)) = next_if_then.next() {
+            // always assume the type of `if`` expr is the same with the `then`` expr
+            (
+                ScalarExpr::If {
+                    cond: Box::new(cond.0),
+                    then: Box::new(then.0),
+                    els: Box::new(build_if_then(next_if_then, els).0),
+                },
+                then.1,
+            )
+        } else {
+            els
+        }
+    }
+    let expr_if = build_if_then(ifs.into_iter(), els);
+    Ok(expr_if)
+}
 /// Convert Substrait Rex into Flow's ScalarExpr
 pub fn from_substrait_rex(
     e: &Expression,
     input_schema: &RelationType,
-    extensions: &HashMap<u32, &String>,
+    extensions: &FunctionExtensions,
 ) -> Result<(ScalarExpr, ColumnType), Error> {
     match &e.rex_type {
         Some(RexType::Literal(lit)) => {
@@ -573,7 +663,9 @@ pub fn from_substrait_rex(
             ))
         }
         Some(RexType::SingularOrList(s)) => {
-            let substrait_expr = s.value.as_ref().unwrap();
+            let substrait_expr = s.value.as_ref().with_context(|| InvalidQuerySnafu {
+                reason: "SingularOrList expression without value",
+            })?;
             // Note that we didn't impl support to in list expr
             if !s.options.is_empty() {
                 return not_impl_err!("In list expression is not supported");
@@ -600,55 +692,17 @@ pub fn from_substrait_rex(
         },
         Some(RexType::ScalarFunction(f)) => from_substrait_scalar_func(f, input_schema, extensions),
         Some(RexType::IfThen(if_then)) => {
-            let ifs: Vec<_> = if_then
-                .ifs
-                .iter()
-                .map(|if_clause| {
-                    let proto_if = if_clause.r#if.as_ref().unwrap();
-                    let proto_then = if_clause.then.as_ref().unwrap();
-                    let cond = from_substrait_rex(proto_if, input_schema, extensions)?;
-                    let then = from_substrait_rex(proto_then, input_schema, extensions)?;
-                    Ok((cond, then))
-                })
-                .try_collect()?;
-            // if no else is presented
-            let els = if_then
-                .r#else
-                .as_ref()
-                .map(|e| from_substrait_rex(e, input_schema, extensions))
-                .transpose()?
-                .unwrap_or_else(|| {
-                    (
-                        ScalarExpr::literal_null(),
-                        ColumnType::new_nullable(CDT::null_datatype()),
-                    )
-                });
-            fn build_if_then(
-                mut next_if_then: impl Iterator<
-                    Item = ((ScalarExpr, ColumnType), (ScalarExpr, ColumnType)),
-                >,
-                els: (ScalarExpr, ColumnType),
-            ) -> (ScalarExpr, ColumnType) {
-                if let Some((cond, then)) = next_if_then.next() {
-                    // always assume the type of `if`` expr is the same with the `then`` expr
-                    (
-                        ScalarExpr::If {
-                            cond: Box::new(cond.0),
-                            then: Box::new(then.0),
-                            els: Box::new(build_if_then(next_if_then, els).0),
-                        },
-                        then.1,
-                    )
-                } else {
-                    els
-                }
-            }
-            let expr_if = build_if_then(ifs.into_iter(), els);
-            Ok(expr_if)
+            from_substrait_ifthen_rex(if_then, input_schema, extensions)
         }
         Some(RexType::Cast(cast)) => {
-            let input = from_substrait_rex(cast.input.as_ref().unwrap(), input_schema, extensions)?;
-            let cast_type = from_substrait_type(cast.r#type.as_ref().unwrap())?;
+            let input = cast.input.as_ref().with_context(|| InvalidQuerySnafu {
+                reason: "Cast expression without input",
+            })?;
+            let input = from_substrait_rex(input, input_schema, extensions)?;
+            let cast_type =
+                from_substrait_type(cast.r#type.as_ref().with_context(|| InvalidQuerySnafu {
+                    reason: "Cast expression without type",
+                })?)?;
             let func = UnaryFunc::from_str_and_type("cast", Some(cast_type.clone()))?;
             Ok((
                 input.0.call_unary(func),
@@ -662,149 +716,6 @@ pub fn from_substrait_rex(
         }
         .fail(),
         _ => not_impl_err!("unsupported rex_type"),
-    }
-}
-
-/// Convert Substrait Expressions to DataFusion Exprs
-pub async fn from_substrait_rex_vec(
-    exprs: &Vec<Expression>,
-    input_schema: &RelationType,
-    extensions: &HashMap<u32, &String>,
-) -> Result<Vec<(ScalarExpr, ColumnType)>, Error> {
-    let mut expressions: Vec<_> = vec![];
-    for expr in exprs {
-        let expression = from_substrait_rex(expr, input_schema, extensions)?;
-        expressions.push(expression);
-    }
-    Ok(expressions)
-}
-
-/// Convert a Substrait literal into a Value and its ConcreteDataType (So that we can know type even if the value is null)
-pub fn from_substrait_literal(lit: &Literal) -> Result<(Value, CDT), Error> {
-    let scalar_value = match &lit.literal_type {
-        Some(LiteralType::Boolean(b)) => (Value::from(*b), CDT::boolean_datatype()),
-        Some(LiteralType::I8(n)) => match lit.type_variation_reference {
-            DEFAULT_TYPE_REF => (Value::from(*n as i8), CDT::int8_datatype()),
-            UNSIGNED_INTEGER_TYPE_REF => (Value::from(*n as u8), CDT::uint8_datatype()),
-            others => not_impl_err!("Unknown type variation reference {others}",)?,
-        },
-        Some(LiteralType::I16(n)) => match lit.type_variation_reference {
-            DEFAULT_TYPE_REF => (Value::from(*n as i16), CDT::int16_datatype()),
-            UNSIGNED_INTEGER_TYPE_REF => (Value::from(*n as u16), CDT::uint16_datatype()),
-            others => not_impl_err!("Unknown type variation reference {others}",)?,
-        },
-        Some(LiteralType::I32(n)) => match lit.type_variation_reference {
-            DEFAULT_TYPE_REF => (Value::from(*n), CDT::int32_datatype()),
-            UNSIGNED_INTEGER_TYPE_REF => (Value::from(*n as u32), CDT::uint32_datatype()),
-            others => not_impl_err!("Unknown type variation reference {others}",)?,
-        },
-        Some(LiteralType::I64(n)) => match lit.type_variation_reference {
-            DEFAULT_TYPE_REF => (Value::from(*n), CDT::int64_datatype()),
-            UNSIGNED_INTEGER_TYPE_REF => (Value::from(*n as u64), CDT::uint64_datatype()),
-            others => not_impl_err!("Unknown type variation reference {others}",)?,
-        },
-        Some(LiteralType::Fp32(f)) => (Value::from(*f), CDT::float32_datatype()),
-        Some(LiteralType::Fp64(f)) => (Value::from(*f), CDT::float64_datatype()),
-        Some(LiteralType::Timestamp(t)) => match lit.type_variation_reference {
-            TIMESTAMP_SECOND_TYPE_REF => (
-                Value::from(Timestamp::new_second(*t)),
-                CDT::timestamp_second_datatype(),
-            ),
-            TIMESTAMP_MILLI_TYPE_REF => (
-                Value::from(Timestamp::new_millisecond(*t)),
-                CDT::timestamp_millisecond_datatype(),
-            ),
-            TIMESTAMP_MICRO_TYPE_REF => (
-                Value::from(Timestamp::new_microsecond(*t)),
-                CDT::timestamp_microsecond_datatype(),
-            ),
-            TIMESTAMP_NANO_TYPE_REF => (
-                Value::from(Timestamp::new_nanosecond(*t)),
-                CDT::timestamp_nanosecond_datatype(),
-            ),
-            others => not_impl_err!("Unknown type variation reference {others}",)?,
-        },
-        Some(LiteralType::Date(d)) => (Value::from(Date::new(*d)), CDT::date_datatype()),
-        Some(LiteralType::String(s)) => (Value::from(s.clone()), CDT::string_datatype()),
-        Some(LiteralType::Binary(b)) | Some(LiteralType::FixedBinary(b)) => {
-            (Value::from(b.clone()), CDT::binary_datatype())
-        }
-        Some(LiteralType::Decimal(d)) => {
-            let value: [u8; 16] = d.value.clone().try_into().map_err(|e| {
-                PlanSnafu {
-                    reason: format!("Failed to parse decimal value from {e:?}"),
-                }
-                .build()
-            })?;
-            let p: u8 = d.precision.try_into().map_err(|e| {
-                PlanSnafu {
-                    reason: format!("Failed to parse decimal precision: {e}"),
-                }
-                .build()
-            })?;
-            let s: i8 = d.scale.try_into().map_err(|e| {
-                PlanSnafu {
-                    reason: format!("Failed to parse decimal scale: {e}"),
-                }
-                .build()
-            })?;
-            let value = i128::from_le_bytes(value);
-            (
-                Value::from(Decimal128::new(value, p, s)),
-                CDT::decimal128_datatype(p, s),
-            )
-        }
-        Some(LiteralType::Null(ntype)) => (Value::Null, from_substrait_type(ntype)?),
-        _ => not_impl_err!("unsupported literal_type")?,
-    };
-    Ok(scalar_value)
-}
-
-fn from_substrait_type(null_type: &substrait::substrait_proto::proto::Type) -> Result<CDT, Error> {
-    if let Some(kind) = &null_type.kind {
-        match kind {
-            Kind::Bool(_) => Ok(CDT::boolean_datatype()),
-            Kind::I8(integer) => match integer.type_variation_reference {
-                DEFAULT_TYPE_REF => Ok(CDT::int8_datatype()),
-                UNSIGNED_INTEGER_TYPE_REF => Ok(CDT::uint8_datatype()),
-                v => not_impl_err!("Unsupported Substrait type variation {v} of type {kind:?}"),
-            },
-            Kind::I16(integer) => match integer.type_variation_reference {
-                DEFAULT_TYPE_REF => Ok(CDT::int16_datatype()),
-                UNSIGNED_INTEGER_TYPE_REF => Ok(CDT::uint16_datatype()),
-                v => not_impl_err!("Unsupported Substrait type variation {v} of type {kind:?}"),
-            },
-            Kind::I32(integer) => match integer.type_variation_reference {
-                DEFAULT_TYPE_REF => Ok(CDT::int32_datatype()),
-                UNSIGNED_INTEGER_TYPE_REF => Ok(CDT::uint32_datatype()),
-                v => not_impl_err!("Unsupported Substrait type variation {v} of type {kind:?}"),
-            },
-            Kind::I64(integer) => match integer.type_variation_reference {
-                DEFAULT_TYPE_REF => Ok(CDT::int64_datatype()),
-                UNSIGNED_INTEGER_TYPE_REF => Ok(CDT::uint64_datatype()),
-                v => not_impl_err!("Unsupported Substrait type variation {v} of type {kind:?}"),
-            },
-            Kind::Fp32(_) => Ok(CDT::float32_datatype()),
-            Kind::Fp64(_) => Ok(CDT::float64_datatype()),
-            Kind::Timestamp(ts) => match ts.type_variation_reference {
-                TIMESTAMP_SECOND_TYPE_REF => Ok(CDT::timestamp_second_datatype()),
-                TIMESTAMP_MILLI_TYPE_REF => Ok(CDT::timestamp_millisecond_datatype()),
-                TIMESTAMP_MICRO_TYPE_REF => Ok(CDT::timestamp_microsecond_datatype()),
-                TIMESTAMP_NANO_TYPE_REF => Ok(CDT::timestamp_nanosecond_datatype()),
-                v => not_impl_err!("Unsupported Substrait type variation {v} of type {kind:?}"),
-            },
-            Kind::Date(date) => match date.type_variation_reference {
-                DATE_32_TYPE_REF => Ok(CDT::date_datatype()),
-                DATE_64_TYPE_REF => Ok(CDT::date_datatype()),
-                v => not_impl_err!("Unsupported Substrait type variation {v} of type {kind:?}"),
-            },
-            Kind::Binary(_) => Ok(CDT::binary_datatype()),
-            Kind::String(_) => Ok(CDT::string_datatype()),
-            Kind::Decimal(d) => Ok(CDT::decimal128_datatype(d.precision as u8, d.scale as i8)),
-            _ => not_impl_err!("Unsupported Substrait type: {kind:?}"),
-        }
-    } else {
-        not_impl_err!("Null type without kind is not supported")
     }
 }
